# go/docs/swagger.yaml
basePath: /
definitions:
...
  enums.SubmissionStatus:
    enum:
    - pending
    - running
    - error
    - completed
    type: string
    x-enum-varnames:
    - SubmissionPending
    - SubmissionRunning
    - SubmissionError
    - SubmissionCompleted
...
  models.Submission:
    properties:
      code:
        description: สำหรับ code exercises
        type: string
      error_message:
        type: string
      failed_count:
        type: integer
      file_name:
        description: ชื่อไฟล์ที่ส่ง
        type: string
      file_size:
        description: ขนาดไฟล์
        type: integer
      file_url:
        description: สำหรับ PDF exercises
        type: string
      material_id:
        description: สำหรับ course materials
        type: string
      mime_type:
        description: ประเภทไฟล์
        type: string
      passed_count:
        type: integer
      results:
        items:
          $ref: '#/definitions/models.SubmissionResult'
        type: array
      status:
        $ref: '#/definitions/enums.SubmissionStatus'
      submission_id:
        type: string
      submitted_at:
        type: string
      total_score:
        type: integer
      user_id:
        type: string
    type: object
  models.SubmissionResult:
    properties:
      actual_output:
        items:
          type: integer
        type: array
      createdAt:
        type: string
      error_message:
        type: string
      result_id:
        type: string
      status:
        type: string
      submission_id:
        type: string
      test_case_id:
        type: string
    type: object
...
  response.StandardResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  types.ApprovePDFSubmissionRequest:
    properties:
      comment:
        minLength: 1
        type: string
      score:
        maximum: 100
        minimum: 0
        type: integer
    required:
    - comment
    - score
    type: object
...
  types.RejectPDFSubmissionRequest:
    properties:
      comment:
        minLength: 1
        type: string
    required:
    - comment
    type: object
...
host: localhost:8080
info:
  contact:
    email: 65070209@kmitl.ac.th
    name: API Support
    url: https://github.com/Project-DSView/backend
  description: This is the backend API for DSView application In Services GO
  termsOfService: http://swagger.io/terms/
  title: DSView API Go Service
  version: 1.0.0-alpha
paths:
...
  /api/materials/{material_id}/submissions:
    get:
      description: Get PDF submissions for a material (Teachers/TAs only)
      parameters:
      - description: Material ID
        in: path
        name: material_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Submission'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get PDF submissions
      tags:
      - pdf-exercises
  /api/materials/{material_id}/submit:
    post:
      consumes:
      - application/json
      description: Submit code for a material-based exercise (new unified system)
      parameters:
      - description: Material ID
        in: path
        name: material_id
        required: true
        type: string
      - description: Code submission
        in: body
        name: request
        required: true
        schema:
          properties:
            code:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                properties:
                  failedCount:
                    type: integer
                  passedCount:
                    type: integer
                  results:
                    items:
                      properties:
                        actualOutput:
                          type: object
                        errorMessage:
                          type: string
                        resultId:
                          type: string
                        status:
                          type: string
                        testCaseId:
                          type: string
                      type: object
                    type: array
                  submissionId:
                    type: string
                  totalScore:
                    type: integer
                type: object
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Submit material exercise
      tags:
      - submissions
...
  /api/submissions/{id}:
    get:
      description: ดูรายละเอียดการส่งงาน (เจ้าของ, Teachers, หรือ TAs ที่ enroll ในคอร์สเท่านั้น)
      parameters:
      - description: Submission ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  code:
                    type: string
                  errorMessage:
                    type: string
                  exerciseId:
                    type: string
                  failedCount:
                    type: integer
                  passedCount:
                    type: integer
                  results:
                    items:
                      properties:
                        actualOutput:
                          type: object
                        errorMessage:
                          type: string
                        resultId:
                          type: string
                        status:
                          type: string
                        testCaseId:
                          type: string
                      type: object
                    type: array
                  status:
                    type: string
                  submissionId:
                    type: string
                  submittedAt:
                    type: string
                  totalScore:
                    type: integer
                  userId:
                    type: string
                type: object
              message:
                type: string
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Submission not found
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Get submission detail
      tags:
      - submissions
  /api/submissions/{submission_id}:
    get:
      description: Get a specific PDF submission
      parameters:
      - description: Submission ID
        in: path
        name: submission_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Submission'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get PDF submission
      tags:
      - pdf-exercises
  /api/submissions/{submission_id}/approve:
    post:
      consumes:
      - application/json
      description: Approve a PDF submission and assign score (Teachers/TAs only)
      parameters:
      - description: Submission ID
        in: path
        name: submission_id
        required: true
        type: string
      - description: Approval data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.ApprovePDFSubmissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Approve PDF submission
      tags:
      - pdf-exercises
  /api/submissions/{submission_id}/reject:
    post:
      consumes:
      - application/json
      description: Reject a PDF submission (Teachers/TAs only)
      parameters:
      - description: Submission ID
        in: path
        name: submission_id
        required: true
        type: string
      - description: Rejection data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.RejectPDFSubmissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Reject PDF submission
      tags:
      - pdf-exercises
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication for secure endpoints.
    in: header
    name: dsview-api-key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"